Adds express server and string generation
